requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==5.2.0
phonenumbers==8.12.31
djangorestframework==3.12.4
requests==2.26.0
geopy==2.2.0
